[refactor] relocate toast files
