--- src/features/toast/components/index.ts
@@ -1,2 +1,0 @@
-export { default as ToastContainer } from "./toastContainer";
-export { default as Toast } from "./toast";
