+++ src/features/toast/index.ts
@@ -1,2 +1,2 @@
-export { useToastStore } from "./toastStore";
-export * from "./components";
+export { default as ToastContainer } from "./containers/toastContainer";
+export { default as useToastStore } from "./stores/toastStore";
